+++ examples/node/basic_tls.ts
@@ -3,7 +3,7 @@ import fs from 'fs'
 
 void (async () => {
   const client = createClient({
-    host: 'https://server.clickhouseconnect.test:8443',
+    url: 'https://server.clickhouseconnect.test:8443',
     tls: {
       ca_cert: fs.readFileSync(
         '../.docker/clickhouse/single_node_tls/certificates/ca.crt'