+++ packages/client-node/src/connection/index.ts
@@ -1,3 +1,4 @@
 export * from './node_base_connection'
 export * from './node_http_connection'
 export * from './node_https_connection'
+export * from './create_connection'